requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
groq
python-dotenv
requests